+++ .vscode/settings.json
@@ -7,6 +7,5 @@
     },
     "[scss]": {
         "editor.tabSize": 2
-    },
-    "files.associations": {}
+    }
 }
