+++ src/cpp/container/array/radixheap_simple.cpp
@@ -11,7 +11,6 @@
 // %require
 // ```
 #include <vector>
-#include <cstdint>
 #include <cassert>
 #include <type_traits>
 #include <algorithm>
@@ -26,7 +25,7 @@ using namespace std;
 
 class RadixHeap {
  public:
-  using value_type = uint64_t;
+  using value_type = long long;
 
  private:
   value_type last_;
