+++ src/cpp/container/likermq/rmq_sparsetable.cpp
@@ -6,7 +6,6 @@
 // - 区間の最小値を計算する．
 // 0-indexedで，[begin,end)．beginを含み，endを含まない．
 // 何度も加減算を繰り返すと内部でオーバーフローを起こす可能性．
-// インデックスは size_t ではなく int で扱う．[TODO]
 // O(log^2N)．最大値のみの機能なら出来そう．[TODO]
 //
 // %usage
