+++ src/cpp/container/likermq/segmenttree/hatopoppo.cpp
@@ -7,7 +7,6 @@
 // - 区間の要素をすべて0にする．
 // - 1になっている区間の要素の個数を求める．
 // 0-indexedで，[begin,end)．beginを含み，endを含まない．
-// インデックスは size_t ではなく int で扱う．[TODO]
 // 反転クエリも書ける。
 //
 // %usage
